### 2023-08-13 #600 @BroVic
Submission of naijR 0.6.1 for review under the rOpenSci Champions Program
Submitting Author Name: Victor Ordu
Submitting Author Github Handle: @BroVic

Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/BroVic/naijR

Version submitted: 0.6.1

Submission type: Standard

Version accepted: TBD

---

```
Package: naijR
Title: Operations to Ease Data Analyses Specific to Nigeria
Version: 0.6.1
    RColorBrewer (>= 1.1.2),
    cli (>= 3.6.0),
    lifecycle (>= 0.2.0),
    grDevices,
    graphics,
    mapdata (>= 2.3.0),
    maps (>= 3.3.0),
    rlang (>= 0.4.0),
    sf (>= 1.0.13),
    stringi (>= 1.7.6),
    utils
    ISOcodes,
    covr,
    here,
    readxl,
    usethis
Authors@R: person("Victor", "Ordu",
              email = "victorordu@outlook.com",
Description: A set of convenience functions as well as geographical/political
  data about Nigeria, aimed at simplifying work with data and information that
  are specific to the country.
RdMacros: lifecycle
```

naijR is an R package that contains a suite of functions that make it easier to deal with data and datasets about the country Nigeria and its constituent parts.

Anyone needing Nigerian geospatial data.

Not to my knowledge.

Yes
