### 2023-08-14 afcb498
Update documentation
Added new examples for some functions and documented return value for fix_region_manual. Also added license to comments of files.
