+++ R/fixreg.R
@@ -1,5 +1,7 @@
 # Source file: fixreg.R
 #
+# GPL-3 License
+#
 # Copyright (C) 2019-2023 Victor Ordu.
 
 #' Fix Region Names
@@ -183,255 +185,8 @@ fix_region.lgas <-
         `states()` or `lgas()` for a more reliable fix"
     )
   
-  invisible(as.character(fix_region(region)))
+  invisible(as.character(fix_region(region, ...)))
 }
-
-
-
-
-  
-## Approximately matches a string on a list of regions i.e States or LGAs.
-## @param str A string to be matched
-## @param regions The values to be matched against, specifically from regions
-##
-## @return A vector of the same length as str with the approximately
-## matched value, which in the case of misspelling should be the correct one.
-##
-## This function is mapped to a vector of states/LGAs
-.fix_region_internal <- function(x, region, interactive = FALSE)
-{
-  stopifnot(is.character(x), is.character(region))
-  cant.fix <- character()
-  fix.status <- character()
-  
-  ## Internal function to enable identification of entries that need to
-  ## be fixed and preparing attributes that will enable further processing
-  ## downstream.
-  get_proper_value <- function(str, regions) {
-    abbrFCT <- .fct_options("abbrev")
-    
-    if (!is.na(match(str, regions)))
-      return(str)
-    
-    if (inherits(regions, "states")) {
-      
-      if (agrepl(str, abbrFCT, max.distance = .pkgLevDistance())
-          && identical(toupper(str), abbrFCT))
-        
-        return(abbrFCT)
-    }
-    
-    ## First remove spaces around slashes and hyphens
-    ## Note: run `.__why_no_pipe()` for rationale behind this approach
-    str <- gsub("\\s\\/", "/", str)
-    str <- gsub("\\/\\s", "/", str)
-    str <- sub("-\\s", "-", str)
-    str <- sub("^Egbado/", "", str) ## TODO: Address hard-coding
-    
-    ## Now, check for exact matching.
-    rgx <- paste0('^', str, '$')
-    good <- unique(grep(rgx, regions, value = TRUE, ignore.case = TRUE))
-    
-    if (length(good) == 1L) 
-      return(good)
-    
-    ## Otherwise check for approximate matches.
-    fixed <- agrep(str, regions, value = TRUE, max.distance = 1)
-    numFixed <- length(fixed)
-    
-    if (numFixed == 1L) {
-      fs <- c(fix.status, fixed)
-      names(fs) <- c(names(fix.status), str)
-      fix.status <<- fs
-      return(fixed)
-    }
-    
-    if (numFixed > 1L && !interactive) {
-      multimatch <- paste(fixed, collapse = ", ")
-      
-      cli::cli_inform(
-        "'{str}' approximately matched more than one region - {multimatch}"
-      )
-    }
-    
-    # if we get to this point, return the misspelt string unchanged
-    cant.fix <<- c(cant.fix, str)
-    str
-  }
-  
-  spellchecked <-
-    vapply(x,
-           get_proper_value,
-           character(1),
-           regions = region,
-           USE.NAMES = FALSE)
-  
-  attr(spellchecked, "misspelt") <- sort(unique(cant.fix))
-  
-  ## Reduce data for reporting on fixes to only the 
-  ## unique instances i.e. avoid redundant output
-  if (length(fix.status) > 1L) {
-    allfix <- names(fix.status)
-    
-    if (anyDuplicated(allfix)) {
-      dups <- which(duplicated(allfix))
[…]
