### 2023-08-14 03e738b
Update NEWS.md
