### 2023-08-14 ffca71b
Move more internal functions to separate files
