+++ R/helpers.R
@@ -16,11 +16,6 @@ globalVariables("states_nigeria")
 # which doesn't belong to any zone is denoted
 # by its own abbreviation and its element is of
 # length 1L.
-stateList <- function()
-{
-  data("states_nigeria", package = "naijR", envir = environment())
-  with(states_nigeria, split(state, gpz))
-}
 
 
 
@@ -67,7 +62,7 @@ country_name <- function()
 
 
 
-.arg_str <- function(arg)
+arg_str <- function(arg)
 {
   deparse(substitute(arg))
 }
@@ -75,7 +70,7 @@ country_name <- function()
 
 
 
-.first_elem_warn <- function(arg)
+first_elem_warn <- function(arg)
 {
   stopifnot(exprs = {is.character(arg) && length(arg) == 1L})
   sprintf("Only the first element of '%s' was used", arg)
@@ -96,4 +91,15 @@ country_name <- function()
 
 .deprec_msg <- function(arg) {
   sprintf("map_ng(%s = )", deparse(substitute(arg)))
+}
+
+
+
+
+# Checks whether a logical argument is correctly passed
+assert.lgl.arg <- function(arg) {
+  argname <- deparse(substitute(arg))
+  
+  if (!is.logical(arg) || is.na(arg))
+    cli_abort("'{argname}' should be TRUE/FALSE")
 }
