+++ R/map.R
@@ -124,7 +124,7 @@ map_ng <- function(region = character(),
                    ...)
 {    ## TODO: Allow this function to accept a matrix e.g. for plotting points
   if (!is.character(region)) {
-    msg <- sprintf("Expected a character vector as '%s'.", .arg_str(region))
+    msg <- sprintf("Expected a character vector as '%s'.", arg_str(region))
     
     addmsg <- if (is.data.frame(region))
       "A data frame was passed; did you mean to use 'data' instead?"
@@ -134,7 +134,7 @@ map_ng <- function(region = character(),
   
   if (!is.null(data) && !is.data.frame(data))
     cli_abort(sprintf("A non-NULL input for '%s' must be a data frame",
-                 .arg_str(data)))
+                 arg_str(data)))
   
   if (is.data.frame(data) && ncol(data) < 2L)
     cli_abort(
@@ -142,11 +142,11 @@ map_ng <- function(region = character(),
     )
   
   if (!is.logical(show.neighbours))
-    cli_abort("'{.arg_str(show.neighbours))}' should be a logical value")
+    cli_abort("'{arg_str(show.neighbours)}' should be a logical value")
   
   if (length(show.neighbours) > 1L) {
-    cli_warn("{.first_elem_warn(.arg_str(show.neighbours))}")
     show.neighbours <- show.neighbours[1]
+    cli_warn("{first_elem_warn(arg_str(show.neighbours))}")
   }
   
   if (show.neighbours)
@@ -222,8 +222,8 @@ map_ng <- function(region = character(),
     
     if (inherits(region, "regions")) {
       rtype <- sub("(.+)(s$)", "\\1", class(region)[1])
-      nmf <- get(paste0("shp.", rtype))$namefield
-      txt <- df.only[[nmf]]
+      namefield <- get(paste0("shp.", rtype))$namefield
+      txt <- df.only[[namefield]]
       
       if (all(is_state(region))) 
         txt <- sub(.fct_options("full"), .fct_options("abbrev"), txt)
@@ -258,7 +258,6 @@ map_ng <- function(region = character(),
       lifecycle::deprecate_warn(.next_minor_version(), .deprec_msg(leg.y))
     
     if (missing(leg.title)) {  # TODO: Change this construct.
-      
       leg.title <- if (is_null(data))
         deparse(substitute(x))
       else
