+++ NAMESPACE
@@ -42,6 +42,7 @@ importFrom(rlang,is_double)
 importFrom(rlang,is_null)
 importFrom(rlang,is_symbol)
 importFrom(rlang,warn)
+importFrom(stats,na.exclude)
 importFrom(utils,data)
 importFrom(utils,head)
 importFrom(utils,tail)
