### 2023-08-17 74bad0a
Update documentation
Added a new section to the help file for `map_ng`.
