### 2023-08-17 9473df9
Refactor to address results from linter check
