+++ R/map.R
@@ -35,12 +35,9 @@ globalVariables(c("STATE", "shp.state", "shp.lga"))
 #' @param legend.text Logical (whether to show the legend) or character vector
 #' (actual strings for the legend). The latter will override whatever is 
 #' provided by \code{categories}, giving the user additional control.
-#' @param leg.x,leg.y Numeric. Position of the legend (deprecated).
 #' @param leg.title String. The legend title. If missing, a default value is
 #' acquired from the data. To turn off the legend title, pass \code{NULL}.
 #' @param plot Logical. Turn actual plotting of the map off or on.
-#' @param leg.orient The orientation of the legend i.e. whether horizontal or
-#' vertical (deprecated). 
 #' @param ... Further arguments passed to \code{\link[sf]{plot}}
 #' 
 #' @details The default value for \code{region} is to print all State 
@@ -118,11 +115,8 @@ map_ng <- function(region = character(),
                    show.neighbours = FALSE,
                    show.text = FALSE,
                    legend.text = NULL,
-                   leg.x = deprecated(),
-                   leg.y = deprecated(),
                    leg.title,
                    plot = TRUE,
-                   leg.orient = deprecated(),
                    ...)
 {    ## TODO: Allow this function to accept a matrix e.g. for plotting points
   if (!is.character(region)) {
@@ -152,9 +146,10 @@ map_ng <- function(region = character(),
   }
   
   if (show.neighbours)
-    cli::cli_alert("Display of neighbouring regions is temporarily disabled")
+    cli::cli_abort("Display of neighbouring regions is temporarily disabled")
   
   region <- .process_region_params(region, call = caller_env())
+  legend.params <- .set_legend_params(legend.text)
   
   value.x <- if (is_null(data) && !is_null(x))
     enquo(x) 
@@ -178,7 +173,8 @@ map_ng <- function(region = character(),
   
   if (use.choropleth) {
     mapq <- expr(.mymap(mapdata, region, plot = plot))
-    cParams <- list(
+    
+    cpleth.inputs <- list(
       region = region,
       value = value.x,
       breaks = breaks,
@@ -195,16 +191,28 @@ map_ng <- function(region = character(),
       else
         as_name(value.x)
       
-      cParams$value <-  data[[value.x]]
-      cParams$region <- data[[region.col]]
+      cpleth.inputs$value <-  data[[value.x]]
+      cpleth.inputs$region <- data[[region.col]]
     }
     
-    cOpts <-
-      .prep_choropleth_opts(mapdata, cParams, dots$col, excluded, exclude.fill)
-    mapq$col <- cOpts$colors
+    cpleth.opts <- .prep_choropleth_opts(
+      mapdata,
+      cpleth.inputs,
+      dots$col,
+      excluded,
+      exclude.fill
+    )
+    mapq$col <- cpleth.opts$colors
     
-    if (lifecycle::is_present(leg.orient))
-      lifecycle::deprecate_warn(.next_minor_version(), .deprec_msg(leg.orient))
+    if (is_null(categories))
+      categories <- cpleth.opts$bins
+    
+    if (is.character(legend.params$text)) {
+      if (length(categories) != length(legend.params$text))
+        cli_abort("Lengths of 'categories' and provided legend do not match")
+      
+      categories <- legend.params$text
+    }
   }
   
   tryCatch({
@@ -215,71 +223,46 @@ map_ng <- function(region = character(),
   if (!is_null(y) && !.xy_within_bounds(sfdata, x, y))
     cli_abort("Coordinates are beyond the bounds of the plotted area")
[…]
