+++ tests/testthat/test-map.R
@@ -5,6 +5,7 @@
 # Copyright (c) 2020-2021 Victor Ordu
 maptype <- "sf"
 
+# Input checks ----
 test_that("Input is validated", {
   myerr1 <- "One or more elements of 'region' is not a Nigerian region"
   myerr2 <- "Expected a character vector as 'region'"
@@ -17,12 +18,12 @@ test_that("Input is validated", {
   expect_error(map_ng(pi), myerr2)
   # expect_warning(map_ng(plot = FALSE, show.neighbours = c(TRUE, TRUE)),
   #                "Only the first element of 'show.neighbours' was used")
-  expect_message(map_ng(plot = FALSE, show.neighbours = TRUE), 
+  expect_error(map_ng(plot = FALSE, show.neighbours = TRUE), 
                  "Display of neighbouring regions is temporarily disabled")
   
-  expect_warning(suppressMessages(map_ng(
+  expect_warning(try(map_ng(
     plot = FALSE, show.neighbours = c(TRUE, FALSE)
-  )),
+  ), silent = TRUE),
   "Only the first element of 'show.neighbours' was used")
   
   expect_error(map_ng(plot = FALSE, show.neighbours = 42),
@@ -35,6 +36,14 @@ test_that("Input is validated", {
   # TODO: Add test case for choropleths with too few regions
 })
 
+test_that("Parameters passed via ellipsis work seamlessly", {
+  expect_s3_class(map_ng(lwd = 2, plot = FALSE), maptype)
+  expect_s3_class(map_ng(lwd = 2, col = 2, plot = FALSE), maptype)
+  expect_error(map_ng(NULL, lwd = 2, col = 2, plot = FALSE), 
+               "Expected a character vector as 'region'")
+})
+
+# Basic maps ----
 test_that("National outline map is plotted", {
   expect_s3_class(map_ng("Nigeria", plot = FALSE), maptype)
 })
@@ -52,6 +61,53 @@ test_that("LGAs are plotted", {
   expect_s3_class(map_ng(lgas("Owerri North"), plot = FALSE), maptype)
 })
 
+# Plain individual sub-national maps ----
+test_that("All individual plain State maps can be drawn", {
+  for (s in states())
+    expect_s3_class(map_ng(s, plot = FALSE), maptype)
+})
+
+test_that("All LGAs within a given State are drawn", {
+  for (s in states()) {
+    expect_s3_class(map_ng(lgas(s), plot = FALSE), maptype)
+  }
+})
+
+test_that("All individual LGA maps can be drawn", {
+  for (s in states()) {
+    lgs <- lgas(s)
+    
+    for (lg in lgs) {
+      x <- suppressWarnings(lgas(lg))
+      state <- attr(x, "State")
+      
+      if (length(state) > 1L)
+        x <- disambiguate_lga(x, state = s)
+      
+      expect_s3_class(map_ng(x, plot = FALSE), maptype)
+    }
+  }
+})
+
+test_that("Map LGAs together as individual blocs", {
+  abiaLga <- lgas("Abia")
+  testMap <- "data/test-map.png"
+  
+  if (file.exists(testMap))
+    file.remove(testMap)
+  
+  png(testMap)
+  val <- try(map_ng(abiaLga), silent = TRUE)
+  dev.off()
+  
+  expect_s3_class(map_ng(abiaLga, plot = FALSE), maptype)
+  expect_false(inherits(val, "try-error"))
+  expect_true(file.exists(testMap))
+  
+  file.remove(testMap)
+})
+
+# Choropleth map ----
 set.seed(4)
 df <- data.frame(region = states(all = TRUE), value = sample(0:6, 37, TRUE))
 mp <- map_ng(plot = FALSE)
[…]
