### 2023-08-18 36208f5
Make a few final changes to improve coverage
There's a deprecated function that was caught by the coverage review, so
that has now veen removed and the documentation and NAMESPACE updated
accordingly.
