### 2023-08-25 #600 @emilyriederer
- [x] Is the reference index page clear (grouped by topic if necessary)? **References page could be further organized with grouping (e.g. `fix` vs `is` functions), but given reasonable number of functions it is still clear**
- [x] **Tests**: If the package has some interactivity / HTTP / plot production etc. are the tests using [state-of-the-art tooling](https://devguide.ropensci.org/building.html#testing)? **NA**
- [ ] **Contributing information**: Is the documentation for contribution clear enough e.g. tokens for tests, playgrounds? **`CONTRIBUTION.md` file exists. It is not extensive, but given that the package does not require a complicated dev environment it is likely fine. However, information on which branch to clo**
- [x] **Project management**: Are the issue and PR trackers in a good shape, e.g. are there outstanding bugs, is it clear when feature requests are meant to be tackled? **Overall issues / PRs are clean. The only thing I notice is a large number of outstanding branches which are likely "stale" from past PRs. Considering deleting these if no longer needed.**
---

This package is in good shape and seems like it could be a powerful accelerant to a data science workflow. Before preceding with the review, I just notice a few changes to be made:

- Regarding Fit, the package is categorized as Geospatial. However, while the package works with location *names* it does not work with geospatial *data types*. Would "data munging" be a more appropriate category since this supports data cleaning operations?

Additionally, the following dimensions could be improved. However, these are not critical to move forward:

- More organization could be added to the References pages (by modifying the pkgdown yaml) to group like-functions together. However, I think it's sufficiently easy to navigate this page as-is.
- More opinionated content could be added to the `CONTRIBUTING.md` file if the author wants to share standards for what he wishes to see in a contribution, if he prefers an issue submitted to discuss before receiving a PR, etc.
- Stale branches could be deleted to clean up the repo

---
