### 2023-08-25 #600 @emilyriederer
Hi @BroVic ! It's great to meet you and I'm looking forward to working as your editor on this package. I left some comments above. I think we are close to looking for reviewers, but I'd love to discuss the "Fit" question with you first. Could you please review our [categories](https://devguide.ropensci.org/policies.html#package-categories) and let me know if you think this package fits the "Data munging" category more than "geospatial"?
