### 2023-08-26 #600 @BroVic
Hello @emilyriederer, indeed the original intent for the package was not to draw maps but help with data validation and/or transformation. The mapping was really a child of necessity due to the absence of elements for drawing subnational maps in existing packages. Future versions will delve into ward level maps, but the focus is to provide objects that can be consumed by actual geospatial packages. So I agree&mdash;we could recategorize it as you've suggested.
