### 2023-08-26 #600 @emilyriederer
Hi @BroVic - thanks for the reply. I think it makes sense to chose the category the best represents the current scope of the package. If you're aligned, please change the category in your initial post at the top.

After that, I'm comfortable moving to identify reviewers. However, I'd encourage you to check out a few of the other notes in my "Editor comments" above in the meantime.

Thanks!
