### 2023-08-28 97fa7ef
Reorganize functions
The is_* and as_* functions have been reorganized in the documentation,
where each is not part of the main help file e.g. states and lgas.
