### 2023-08-31 993396d
Reorganize the documentation of methods
Most of the methods for internal generics were modified by using the
actual classes representing adminstrative regions, rather than the
abstrat 'regions' class. The motivation was for more clarity in the
documentation.
