### 2023-09-02 #600 @emilyriederer
Hi @BroVic - I'm thrilled to introduce @mcsiple as the first reviewer for your package! Her expertise is in both text and spatial data so she seemed like a great fit for this package. I should have an update on the second reviewer soon.
