### 2023-09-03 #600 @BroVic
Thank you @emilyriederer. I really appreciate your efforts and I can't wait to learn from you guys. Cheers!
