### 2023-09-03 #600 @emilyriederer
Hi @BroVic ! I'm delighted to introduce the @cagrigokcek as the second reviewer. His domain expertise is in political science, so I'm sure that will help him empathize with potential users of this package. Earlier this summer, he was featured in [this rOpenSci blog post](https://ropensci.org/blog/2023/06/19/ropensci-champions-program-teams-meet-alican-cagri-gokcek-and-elio-campitelli/)
