### 2023-09-22 #600 @mcsiple
## `naijR` Package Review
Good work on this package @BroVic et al. I used the rOpenSci review template to set up my checks. My limited comments/suggestions can be found at the bottom of this review. Please let me know if you have any questions about anything I wrote.

- *The vignettes are easy to use and both work. I had to use `devtools::build_vignettes()` after installing directly from github source. The nigeria-maps vignette had a build error because one or two calls to `map_ng()` use `fill` as an argument when I think it should be `exclude_fill`. Suggest adding a `build_vignettes` argument to install instructions in README.md .(`remotes::install_github("BroVic/naijR",build_vignettes = TRUE)`)*
- [ ] **Function Documentation:** for all
`as_lga()` / `is_lga()` -- works and examples run great.
`fix_region()` - I like the interactive option for fixing region names; this is a smart addition to the package.
`map_ng()` - In documentation for map_ng, "sw" is not recognized as a state (i.e., when you run `map_ng(states("sw"))` you get a warning "One or more items is not a State"). I am a little unclear about the CRS for the points in the "Mapping of point data" vignette. I assume that many users will have lats and lons and will want to plot those over the map, or use the spatial data to assign points to states. If there is a way to do this with `naijR`, can you demonstrate in the nigera-maps vignette? Or otherwise provide an example where the `sf` object generated from `map_ng()` is used.

*But see note about vignettes above*
Test coverage is good (I used `covr::package_coverage()`)

---

Great work putting together the package so comprehensively. It made it easier to review and will be easier for users, too. I have just a couple of comments.

- Installs with devtools and from source worked great. See suggestions above about adding a vignette argument.
- You might want to change the dependency for maptools here to avoid the warning messages that users will likely get about maptools being discontinued in favor of sf/terra. Suppressing warnings about this is in your News doc so I bet you are already thinking about doing this.
-`map()` and `regions()` produced warnings when I ran `goodpractice::gp()` about unit test coverage, but I think 97% coverage is pretty good so I will leave it to the editor to decide if additional coverage is really necessarily on those.
- `{goodpractice}` also suggests avoiding `sapply()` (used in admin.R, helpers.R and utils.R).
- An optional thing that would be nice would be some additional detail about how the name disambiguations are produced. For example, when `fix_region()` is automatic, what is the underlying method used to get the disambiguation? This may help others who are working on similar data packages, and help them troubleshoot when misspellings are too 'far gone' to produce an automatic correction.
- Minor note - if there is a source for the geographical data or the "official" anglicized spelling of the LGA or state names, that might be worth adding.
