### 2023-09-29 #600 @emilyriederer
Thanks for the update, @cagrigokcek ! Yes, that should work just fine. I
appreciate your taking the time to do a thorough review!

On Thu, Sep 28, 2023 at 3:43 PM A. Cagri gokcek ***@***.***>
wrote:

> Hi Emily,
>
> Is it ok if I submit my review by the end of Sunday night? I know that I
> already demanded a few days extra. But, this will be the last and final
> request from me. Please let me know.
>
> Best.
>
> Cagri
>
>
>
> From: Emily Riederer ***@***.***>
> Sent: Thursday, September 28, 2023 2:38 AM
> To: ropensci/software-review ***@***.***>
> Cc: A. Cagri gokcek ***@***.***>; Mention ***@***.***>
> Subject: Re: [ropensci/software-review] Submission of naijR 0.6.1 for
> review under the rOpenSci Champions Program (Issue #600)
>
>
>
> Thanks for the review, @mcsiple <https://github.com/mcsiple> ! I really
> appreciate the thorough suggestions here.
>
> @cagrigokcek <https://github.com/cagrigokcek> , do you have an update
> from your side of when you will have your review completed?
>
