### 2023-10-02 #600 @cagrigokcek
I am very happy to give you feedback. As you know, this is my first chance of getting involved in a peer-review process. That said, please forgive me if I miss any crucial point in this review, and don't hesitate to forward me any type of accessibility-related questions. In this review, I majorly focused on your map_ng() function. My main goal was checking the level of accessibility of maps created by your package. Thus, I skipped some items in the check list due to the lack of enough knowledge to test them. Drawing insights from this review process, I also offer an extra bullet point under the heading of functionality to @emilyriederer on behalf of @rOpenSci. I added a new bullet point below, and left my feedback there. I hope that it will help @BroVic to make naijR more inclusive and accessible for everyone.:

- [ ] **Accessibility of package:**

I will recommend to benefit from BrailleR packages' functions for discovering maps in the package. the pertinant information about BrailleR in regards to making data visualization more accessible for the blind can be obtained via this vignette:
https://cran.r-project.org/web/packages/BrailleR/vignettes/ExploringGraphs.html
When I created two objects by map_ng(), I was not able to use BraileR's relevant commends for graphs. At this moment, I should also admit that I am not sure whether BrailleR's commands work for maps. I know that BrailleR works smoothly with ggplot2 and tidyverse in general. I am sharing my codes and the message received in R for @BroVic's consideration.

``` r
library(naijR)
library(BrailleR)
map1=map_ng("Nigeria")

Describe(map1)
#> No description of this type of object has been written at this time.
#>
#> General Description:
#> none yet
#>
#> R hints:
#> none yet

map2=map_ng()

Describe(map2)
#> No description of this type of object has been written at this time.
#>
#> General Description:
#> none yet
#>
#> R hints:
#> none yet
```

Describe function, if it worked, could allow a screen reader user to understand basic features of the maps such as the state names displayed in the map, their geographic position in the map, the color of states, if any, in the map, and any other features considered fundamentally important by the author of the package.
Here, I only demonstrated Describe() function. Other functions of BrailleR (e.g. VI()) can be integrated in naijR package to increase the inclusivity of naijR package.
Overall, as it is seen, a development task for naijR can be making naijR package compatible with BrailleR package's commands related to data visualization.

---

I am very excited to see what kind of changes can ebedone for improving the accessibility of maps in this package. Please don't hesitate to keep in touch with me anytime if you need me to test your new features.
