### 2023-10-04 538bdb4
Address changes requested by reviewers
1. Update the package contributor metadata.
2. Force vignette builds during source installs from GitHub. Additionally, force installation of dependencies for `Imports` and `Suggests`.
3. Remove obsolete argument `fill` for `map_ng()` still found in the vignette.
4. Provided more clarity in the case where a geopolitical zone is passed to the `states` argument of `states()`. This now signals an error with information on how to use it correctly.

**Other Changes:**

- Improvements to function help.
