+++ tests/testthat/test-regions.R
@@ -1,15 +1,15 @@
 # Copyright (C) 2019-2021 Victor Ordu.
-# 
+#
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
-# 
+#
 # This program is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
-# 
+#
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>
 
@@ -23,7 +23,7 @@ test_that("input is validated", {
   gpz.no.char <- "argument 'gpz' is not of type 'character'"
   all.no.lgl <- "'all' should be TRUE/FALSE"
   wrn.no.lgl <- "'warn' should be TRUE/FALSE"
-  
+
   expect_error(states(gpz = 99), gpz.no.char)
   expect_error(states(gpz = NA), gpz.no.char)
   expect_error(states(gpz = "gpz"))
@@ -35,6 +35,8 @@ test_that("input is validated", {
   expect_error(states(warn = "full"), wrn.no.lgl)
   expect_error(states(warn = NA), wrn.no.lgl)
   expect_error(states(warn = NULL), wrn.no.lgl)
+  expect_error(states("sw"),
+               "'sw' represents a geopolitical zone, not a State")
 })
 
 
@@ -43,7 +45,7 @@ test_that("input is validated", {
 test_that("'states' object is properly constructed", {
   wrn <- "One or more items is not a State. Spelling error\\?"
   shuffled <- sample(states(), 37L, FALSE)
-  
+
   expect_warning(states(c("Oyo", "Legos")), wrn)
   expect_silent(states(c("xxx", "Benue"), warn = FALSE))
   expect_warning(states(c("kentucky", "Bornu", "Abia")))
@@ -57,7 +59,7 @@ test_that("FCT can be selectively removed", {
   allstates <- states()
   notall <- states(all = FALSE)
   fct <- "Federal Capital Territory"
-  
+
   expect_length(notall, 36L)
   expect_length(allstates, 37L)
   expect_true(fct %in% allstates)
@@ -69,11 +71,11 @@ test_that("FCT can be selectively removed", {
 
 
 
-## ---- LGA related tests ---- 
+## ---- LGA related tests ----
 
 test_that("illegal input is caught early", {
   chErr <- "Expected an object of type 'character'"
-  
+
   expect_error(lgas("Saarland"))
   expect_error(lgas("Maryland"), "None of the items is a valid LGA")
   expect_error(lgas(888), chErr)
@@ -86,7 +88,7 @@ test_that("illegal input is caught early", {
 test_that("LGAs are returned correctly", {
   res <- lgas("Plateau")
   res2 <- lgas(nam <- c("Borno", "Abia"))
-  
+
   expect_match(res, "Pankshin", all = FALSE)
   expect_length(res, 17L)
   expect_type(res, "character")
@@ -94,18 +96,18 @@ test_that("LGAs are returned correctly", {
   expect_type(res2, "list")
   expect_named(res2, nam)
   expect_length(res2, 2L)
-  expect_warning(lgas(c("Oyo West", "Obomo Ngwa")), 
+  expect_warning(lgas(c("Oyo West", "Obomo Ngwa")),
[…]
