+++ DESCRIPTION
@@ -26,10 +26,27 @@ Suggests:
     rmarkdown,
     testthat (>= 3.0.0),
     usethis
-Authors@R: person("Victor", "Ordu", 
-              email = "victorordu@outlook.com",
-              role = c("aut", "cre"),
-              comment = c(ORCID = "0000-0003-3716-0668"))
+Authors@R: 
+	c(person(given = "Victor", 
+	         family = "Ordu",
+             role = c("aut", "cre"),
+             email = "victorordu@outlook.com",
+             comment = c(ORCID = "0000-0003-3716-0668")),
+	  person(given = "Laura",
+			 family = "DeCicco",
+			 role = "ctb",
+			 email = "ldecicco@usgs.gov"),
+	  person(given = "Emily",
+	         family = "Riederer",
+			 role = "rev",
+			 email = "emilyriederer@gmail.com"),
+	  person(given = "Margaret",
+	         family = "Siple",
+			 role = "rev"),
+	  person(given = "Cagri",
+	         family = "Alican",
+			 role = "rev")
+	)
 Description: A set of convenience functions as well as geographical/political
   data about Nigeria, aimed at simplifying work with data and information that
   are specific to the country.
