### 2023-10-06 #600 @emilyriederer
Thank you so much for the reviews @mcsiple and @cagrigokcek ! Could you please both share an estimated time spent reviewing? This is a metric we track internally and I need it so I can trigger the bot to formally "submit" the reviews
