### 2023-10-06 #600 @cagrigokcek
Hi Emily,

I didn’t actually count my hours while reviewing the package. I spent few hours in two days to complete the review. I should say that I didn’t spend more than 5 hours in two days.

Thanks.

Cagri

From: Emily Riederer ***@***.***>
Sent: Thursday, October 5, 2023 5:03 PM
To: ropensci/software-review ***@***.***>
Cc: A. Cagri gokcek ***@***.***>; Mention ***@***.***>
Subject: Re: [ropensci/software-review] Submission of naijR 0.6.1 for review under the rOpenSci Champions Program (Issue #600)

Thank you so much for the reviews @mcsiple <https://github.com/mcsiple>  and @cagrigokcek <https://github.com/cagrigokcek>  ! Could you please both share an estimated time spent reviewing? This is a metric we track internally and I need it so I can trigger the bot to formally "submit" the reviews
