### 2023-10-15 #600 @emilyriederer
Hi @BroVic ! I hope you are well. I thought I'd drop a line and see how you are feeling about any changes inspired by the review comments? It would be great if we can give our reviewers a heads-up on when to expect a final version to review for approval. Thanks!
