### 2023-10-15 #600 @BroVic
Dear @emilyriederer I thought I would have finished applying the requested updates by now, but I had a little battle with malaria last weekend. I'm fine now and recuperating, so you will definitely hear from me in this regard very soon. Thanks. ❤️‍🩹
