+++ tests/testthat/test-map.R
@@ -416,6 +416,13 @@ test_that("Points are mapped", {
   
   expect_error(map_ng(x = 15, y = 45, plot = FALSE),
                "Coordinates are beyond the bounds of the plotted area")
+  
+  
+  capitals <- read.csv(here::here("data-raw/state-capitals.csv"))
+  out <- map_ng(x = capitals$longitude, y = capitals$latitude)
+  
+  expect_s3_class(out, "sf")
+  
 })
 
 # Labels ----
