### 2023-10-25 7ba2737
Fix bug introcuced in script for making spatial objects
The objects were not being retrieved from the global environment where they were created. This failure happened because the script is now being evaluated in a self-contained environment via `local()`.
