### 2023-11-05 6dc38fa
Update unit tests
Apart from some refactoring, one of the tests that kept producing a plot was identified and modified.
