### 2023-11-16 #600 @BroVic
Hello @mcsiple, thank you very much for your review. It enabled me to effect several improvements and opened my eyes to other issues you didn't mention directly.

I have 2 comments:

1. I am aware that `sapply` is not type-safe, but in this case I would plead that we let it be as it is. The places where it is used are not accessible to the end user - those are utility files for generating some of the package's internal data. I had to use the function because I can return a named object and I made sure it returns a list via the `simplify = FALSE` argument. The other option would have been to use the {purrr}  package, but I don't believe this justified adding the dependency. If I tried, using other functions, everything just broke.
2. Unfortunately, there is no canonical source for the correct versions of the administrative region names. Even the Nigerian Constitution, where the Local Government Areas are listed, overt mistakes as well as contested names! Right now, they are working on a law that addresses these naming discrepancies, and this is one of the things that made me develop this package in the first place. So, it's still evolving and when we have something more concrete, I will definitely implement your suggestion.

Once again, thanks!
