### 2023-11-25 #600 @mcsiple
Hi @BroVic and @emilyriederer , sorry for the delay. I was out of the office for a longer period than usual and am just getting back to my emails and GitHub. @BroVic , I am fine with those two comments. For (1) I agree that it's not worth adding a dependency (and I think `purrr` is sometimes slower than `sapply()`) so that is okay. And for (2) I am impressed at the ways you have found to include this 'naming flexibility' in your package. That is an exciting software challenge indeed  🙂

Anyway, I am satisfied with the changes and your comments.
