### 2023-11-28 #600 @cagrigokcek
Hi Emily, May I request a few days extension because I struggle with a severe tooth pain which decreases my concentration significantly? I promise to complete the form before the weekend.

Best.

Cagri

From: Emily Riederer ***@***.***>
Sent: Sunday, November 26, 2023 5:47 PM
To: ropensci/software-review ***@***.***>
Cc: A. Cagri gokcek ***@***.***>; Mention ***@***.***>
Subject: Re: [ropensci/software-review] Submission of naijR 0.6.1 for review under the rOpenSci Champions Program (Issue #600)

Thanks so much @mcsiple <https://github.com/mcsiple>  ! I hope you had a nice break from the office.

If both you and @cagrigokcek <https://github.com/cagrigokcek>  could please use the reviewer approval form <https://devguide.ropensci.org/approval2template.html>  once you are satisfied, that will be the final step!
