### 2023-12-02 #600 @cagrigokcek
Hi viktor,

I am so sorry to see your message very late. I totally understand your situation about BrailleR package. As you said, it can be a further goal for you to develop packages which can be integrated with available packages related to accessibility in R later. I find your explanation very legitimate. Please don’t hesitate to reach me out when you need a tester in the future about features related to accessibility in R.

Best.

Cagri

From: Victor Ordu ***@***.***>
Sent: Thursday, November 16, 2023 8:46 AM
To: ropensci/software-review ***@***.***>
Cc: A. Cagri gokcek ***@***.***>; Mention ***@***.***>
Subject: Re: [ropensci/software-review] Submission of naijR 0.6.1 for review under the rOpenSci Champions Program (Issue #600)

Dear @cagrigokcek <https://github.com/cagrigokcek> , thanks again for your review. I would like to make a few comments and if you think I am inadequately informed, feel free to let me know:

1.	Although this package has a function that draws maps, this is not its main purpose. The central idea was to provide data that are specific to Nigeria and to help users work better with those data. The map_ng() function returns an sf object, which makes it easier to use with other real geospatial packages.
2.	I took a look at how best one could link the package to {BrailleR} as suggested. My preferred approach would have been to run the output of map_ng() through ggplot::geom_sf() and from there attempt to get textual data from the map. I actually thought that ggplot2 provides methods for working with {BrailleR}. I can now see that methods like BralleR::Describe are defined by that package's author(s). So, I believe the onus seems to lie with them to create methods that can generically provide accessible output from objects created by widely used and popular packages like {ggplot2}

So, in light of the foregoing, I might have to decline this feature for {naijR}, the reason being that it is out of scope (for now). But knowing how important this issue of accessibility is, I will bear it in mind in future development of this package, and if a decision is made to expand the scope, then one might consider taking on {BrailleR} as a dependency.

Cheers.
