### 2023-12-02 #600 @cagrigokcek
Hi Emily and Viktor,

I copy my review approval form below. Please let me know if there is any missing part in the form.

I congratulate Viktor again.

Best.

Cagri

- [X ] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

Estimated hours spent reviewing: 5h30m

From: Emily Riederer ***@***.***>
Sent: Sunday, November 26, 2023 5:47 PM
To: ropensci/software-review ***@***.***>
Cc: A. Cagri gokcek ***@***.***>; Mention ***@***.***>
Subject: Re: [ropensci/software-review] Submission of naijR 0.6.1 for review under the rOpenSci Champions Program (Issue #600)

Thanks so much @mcsiple <https://github.com/mcsiple>  ! I hope you had a nice break from the office.

If both you and @cagrigokcek <https://github.com/cagrigokcek>  could please use the reviewer approval form <https://devguide.ropensci.org/approval2template.html>  once you are satisfied, that will be the final step!
