### 2023-12-02 #600 @emilyriederer
Congratulations @BroVic ! It looks like both reviewers have approved your package. (@mcsiple - would appreciate if you could fill out the last template just to close the loop, but I appreciate that we have your answer!) I will trigger the Approval bot now. Please review the messages it sends carefully with instructions for next steps to onboard this package to the rOpenSci family
