### 2023-12-02 #600 @ropensci-review-bot
Approved! Thanks @BroVic for submitting and @mcsiple, @cagrigokcek for your reviews! :grin:
